Reimplement POST request
